+++ Microsoft.Azure.Cosmos/src/Resource/Container/ContainerCore.Items.cs
@@ -365,7 +365,7 @@ namespace Microsoft.Azure.Cosmos
 
             return new FeedIteratorCore<T>(
                 feedIterator: feedIterator,
-                responseCreator: this.ClientContext.ResponseFactory.CreateQueryFeedResponse<T>);
+                responseCreator: this.ClientContext.ResponseFactory.CreateQueryFeedUserTypeResponse<T>);
         }
 
         public override IOrderedQueryable<T> GetItemLinqQueryable<T>(
@@ -519,7 +519,7 @@ namespace Microsoft.Azure.Cosmos
             RequestOptions requestOptions,
             CancellationToken cancellationToken)
         {
-            Stream streamPayload = this.ClientContext.CosmosSerializer.ToStream<T>(item);
+            Stream streamPayload = this.ClientContext.SerializerCore.ToStream<T>(item);
 
             // User specified PK value, no need to extract it
             if (partitionKey.HasValue)