+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/SettingsContractTests.cs
@@ -395,18 +395,18 @@ namespace Microsoft.Azure.Cosmos.Tests
         {
             string containerJsonString = "{\"indexingPolicy\":{\"automatic\":true,\"indexingMode\":\"Consistent\",\"includedPaths\":[{\"path\":\"/*\",\"indexes\":[{\"dataType\":\"Number\",\"precision\":-1,\"kind\":\"Range\"},{\"dataType\":\"String\",\"precision\":-1,\"kind\":\"Range\"}]}],\"excludedPaths\":[{\"path\":\"/\\\"_etag\\\"/?\"}],\"compositeIndexes\":[],\"spatialIndexes\":[]},\"id\":\"MigrationTest\",\"partitionKey\":{\"paths\":[\"/id\"],\"kind\":\"Hash\"}}";
 
-            CosmosJsonDotNetSerializer cosmosSerializer = new CosmosJsonDotNetSerializer();
+            CosmosJsonDotNetSerializer serializerCore = new CosmosJsonDotNetSerializer();
             ContainerProperties containerProperties = null;
             using (MemoryStream memory = new MemoryStream(Encoding.UTF8.GetBytes(containerJsonString)))
             {
-                containerProperties = cosmosSerializer.FromStream<ContainerProperties>(memory);
+                containerProperties = serializerCore.FromStream<ContainerProperties>(memory);
             }
 
             Assert.IsNotNull(containerProperties);
             Assert.AreEqual("MigrationTest", containerProperties.Id);
 
             string containerJsonAfterConversion = null;
-            using (Stream stream = cosmosSerializer.ToStream<ContainerProperties>(containerProperties))
+            using (Stream stream = serializerCore.ToStream<ContainerProperties>(containerProperties))
             {
                 using (StreamReader sr = new StreamReader(stream))
                 {
@@ -502,13 +502,13 @@ namespace Microsoft.Azure.Cosmos.Tests
                 documentCollection.SaveTo(memoryStream);
                 memoryStream.Position = 0;
 
-                CosmosJsonDotNetSerializer cosmosSerializer = new CosmosJsonDotNetSerializer();
-                ContainerProperties containerProperties = cosmosSerializer.FromStream<ContainerProperties>(memoryStream);
+                CosmosJsonDotNetSerializer serializerCore = new CosmosJsonDotNetSerializer();
+                ContainerProperties containerProperties = serializerCore.FromStream<ContainerProperties>(memoryStream);
 
                 Assert.IsNotNull(containerProperties);
                 Assert.AreEqual(containerId, containerProperties.Id);
 
-                using (Stream stream = cosmosSerializer.ToStream<ContainerProperties>(containerProperties))
+                using (Stream stream = serializerCore.ToStream<ContainerProperties>(containerProperties))
                 {
                     using (StreamReader sr = new StreamReader(stream))
                     {